separated useScrollAttributes into it’s own folder
also moved all tests specific to that hook to that folder as well
