+++ src/components/Hscroll/index.test.js
@@ -2,7 +2,6 @@ import React from 'react';
 import { shallow } from 'enzyme';
 
 import Hscroll from '.';
-import { getScrollAttributes } from './useScrollAttributes';
 
 const WIDTH_CONTAINER = 400;
 const WIDTH_CONTENT = 800;
@@ -31,49 +30,6 @@ describe('Hscroll component', () => {
   it('matches snapshot (enable gutter)', () => {
     const wrapper = renderComponent({ enableGutter: true });
     expect(wrapper).toMatchSnapshot();
-  });
-
-  describe('getScrollAttributes', () => {
-
-    it('returns correct attributes for partially scrolling', () => {
-      const MOCK_EVENT = {
-        target: {
-          scrollLeft: 50,
-          clientWidth: WIDTH_CONTAINER
-        }
-      };
-      const { scrolled, scrollEnd } = getScrollAttributes(MOCK_EVENT, WIDTH_CONTENT);
-
-      expect(scrolled).toBe(true);
-      expect(scrollEnd).toBe(false);
-    });
-
-    it('returns correct attributes for scrolling completely to the end', () => {
-      const MOCK_EVENT = {
-        target: {
-          scrollLeft: WIDTH_CONTAINER,
-          clientWidth: WIDTH_CONTAINER
-        }
-      };
-      const { scrolled, scrollEnd } = getScrollAttributes(MOCK_EVENT, WIDTH_CONTENT);
-
-      expect(scrolled).toBe(true);
-      expect(scrollEnd).toBe(true);
-    });
-
-    it('returns correct attributes for scrolling to the beginning', () => {
-      const MOCK_EVENT = {
-        target: {
-          scrollLeft: 0,
-          clientWidth: WIDTH_CONTAINER
-        }
-      };
-      const { scrolled, scrollEnd } = getScrollAttributes(MOCK_EVENT, WIDTH_CONTENT);
-
-      expect(scrolled).toBe(false);
-      expect(scrollEnd).toBe(false);
-    })
-
   });
 
 });
